Face mask detection
This yolo algorithm is trained to detect if person in image is wearing mask or not. Although the model is trained on only 46 images and for 1000 epochs but it predicts the results quite well.The accuracy of the model can be greatly increased if more data is provided and the number of epochs are increased.